+++ snippets/api/helpers.php
@@ -1,47 +1,52 @@
 <?php
-/*
- * This is an unofficial snippet, with no support! You may copy-paste and use at your risk.
- */
+
+use Illuminate\Http\JsonResponse;
 
 /**
  * Wrap the success result in a JSON response.
  *
  * @param mixed $result
  *
- * @return \Illuminate\Http\JsonResponse
+ * @return JsonResponse
  */
 function api_success($result)
 {
     return response()->json([
-        'success' => true,
+        'status' => 'OK',
         'result'  => $result,
-    ]);
+    ], 200, [], JsonResponse::DEFAULT_ENCODING_OPTIONS | JSON_UNESCAPED_UNICODE);
 }
 
 /**
  * The JSON API error response.
  *
  * @param int $statusCode
+ * @param null|string $errorCode
  * @param null|string $message
- * @param array $errors
+ * @param array $info
  *
- * @return \Illuminate\Http\JsonResponse
+ * @return JsonResponse
  */
-function api_error($statusCode, $message = null, array $errors = [])
+function api_error($statusCode, $errorCode = null, $message = null, array $info = [])
 {
-    return response()->json(api_error_body($statusCode, $message, $errors), $statusCode);
+    return response()->json(
+        api_error_body($statusCode, $errorCode, $message, $info),
+        $statusCode,
+        [],
+        JsonResponse::DEFAULT_ENCODING_OPTIONS | JSON_UNESCAPED_UNICODE);
 }
 
 /**
  * The body of API error response in array format.
  *
  * @param int $statusCode
+ * @param null|string $errorCode
  * @param null|string $message
- * @param array $errors
+ * @param array $info
  *
  * @return array
  */
-function api_error_body($statusCode, $message = null, array $errors = [])
+function api_error_body($statusCode, $errorCode, $message, array $info)
 {
     if (is_null($message)) {
         if (trans('api_errors.'.$statusCode) === 'api_errors.'.$statusCode) {
@@ -52,12 +57,15 @@ function api_error_body($statusCode, $message = null, array $errors = [])
     }
 
     $body = [
-        'success' => false,
-        'message' => $message,
+        'status' => 'ERROR',
+        'error' => [
+            'code' => $errorCode,
+            'message' => $message,
+        ],
     ];
 
-    if ($errors) {
-        $body['errors'] = $errors;
+    if ($info) {
+        $body['error']['info'] = $info;
     }
 
     return $body;